Updates how-to-contribute docs (#9643)
Recently I tried to run the instance of gatsby's website locally. 
By using `gatsby develop` in /www, unfortunately, it got stuck. Refer #9640 for more info.
Thanks to @DSchau he suggested trying `GATSBY_SCREENSHOT_PLACEHOLDER=true` to run the dev server. But for a new contributor, it's hard for him to figure this out. 

So I made some changes in the `how-to-contribute` docs.