docs(browser-apis): fix replaceHydrateFunction JSDoc comment (#9701)
`replaceHydrateFunction` is not currently showing up in the [Gatsby Browser APIs](https://www.gatsbyjs.org/docs/browser-apis/) section of the documentation site. The entries there are generated based on the JSDoc comments in `packages/gatsby/src/utils/api-browser-docs.js`, and the comment for `replaceHydrateFunction` was missing an asterisk, making it invalid JSDoc.