feat(gatsby-source-mongodb): mongodb to ver 3, added ability to pass entire connection string (#11698)
## Questions/Concerns
This PR requires testing. Also, are there any other features of more recent versions of MongoDB or the mongodb npm package that should be incorporated?

## Description

This change allows the use of connections strings compatible with MongoDB 3.6+ by passing in the entire connection string provided by a service such as MongoDB Atlas or mLab. The default configuration is retained for legacy versions and for local defaults.This also updates the mongodb npm package to version 3+.

## Related Issues

None, just my personal need to connect to higher version DB than could connect via the current plugin.