feat(www): don't check for overview frontmatter (#13060)
Fixes `[[guidelist]]` not being replaced in few pages (i.e. in https://www.gatsbyjs.org/docs/winning-over-stakeholders/ )

The code is handling cases when there are no children (tried building locally and it worked just fine).

Follow up to this would be to clear `.md` files that use `overview` and adjust page query. But this is just quick fix.