feat(gatsby-source-drupal): Disable caching + add http/2 agent (#32012)
* feat(gatsby-source-drupal): Force revalidation on Drupal API requests

Talking to Gatsby/Drupal users — most set long max-age for the Drupal cache to
keep their edge http cache as fresh as possible as Drupal can directly purge
its edge http cache.

But this has the unfortunate side-effect with the recent http client change in
https://github.com/gatsbyjs/gatsby/pull/31514 that API calls aren't
revalidated. Meaning that a user could change some content in Drupal and not
see those updates in their Gatsby site until the Drupal cache expires in the
Gatsby cache.

This PR removes the `cache-control` header from Drupal API responses so that
we only can use `etag` for caching which forces revalidation on every request.

* Remove debugging line

* Add http/2 agent

* Disable cache — it's slower than refetch with revalidation

* push don't concat

* consistent return

* Fix statusCode check for got

* Actually this was correct...

* Actually use http/2